code-interpreter/pyproject.toml
[project]
name = "code-interpreter"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain>=1.0.0",
    "langchain-openai>=1.0.0",
    "python-dotenv>=1.0.0",
    "pandas>=2.0.0",
    "qrcode>=7.0.0",
]
